==> code/04_run_models.Rmd <==
---
title: "Fit MRP Models with Mobility Data"
author: Casey Breen
---

## Summary

This notebook contains code to fit two main hierarchical models: one model to estimate non-household contacts and one to model to estimate household contacts.

```{r}
## library here package 
library(here)

## Load helper functions and library all packages used for analysis 
source(here("code", "helpers.R"))

## prepared for model  
df_formodel <- read_rds(here('data', 'input', 'bics_data_waves_1-6.rds')) 
```

Recode stringency index and pca to 3-week rolling average  

```{r}
## top code total number of reported contacts at 30 
top_code <- function(x) {
  x <- case_when(
    x > 30 ~ 30,
    TRUE ~ x
  )
}

## topcode data.frame 
df_formodel_topcode <- df_formodel %>%
  mutate(across(c(starts_with("alter_age")), top_code))

df_formodel <- df_formodel %>% 
  mutate(rolling_pca_3wks_factor = as.factor(ntile(rolling_pca_3wks, 3))) %>% 
  mutate(stringency_index_factor = as.factor(ntile(stringency_index, 3)))
```


```{r}
## filter to respondents in household with more than 1 person 
df_formodel_topcode_household <- df_formodel_topcode %>%
  filter(hhsize != "1") 
```


Run final household model for paper
Note: no mobility data used for within-household contacts 

```{r}
## run model (estimated run time: 978 seconds)
hh_model <- brm(
  formula = mvbind(alter_age_0_18_hh, alter_age_18_25_hh, alter_age_25_35_hh, alter_age_35_45_hh, alter_age_45_55_hh, alter_age_55_65_hh, alter_age_65_100_hh) ~ 
    agecat +
    gender +
    hhsize + 
    race + 
    agecat:gender + 
    (1 | state),
  family = 'negbinomial',
  inits = 0,
  data = df_formodel_topcode_household,
  chains = 4,
  cores = 16,
  iter = 4000)

## write out model 
write_rds(hh_model, file = here('data', 'models', 'hh_model.rds'))
```


Run main model non-hh model with mobility for paper 

```{r}
## mobility + monthly national cubic + state-level hierarchical cubic term 
## Total execution time: 5346.0 seconds.
nonhh_model <- brm(
  formula = mvbind(alter_age_0_18, alter_age_18_25, alter_age_25_35, alter_age_35_45, alter_age_45_55, alter_age_55_65, alter_age_65_100) ~ 
    agecat +
    gender +
    hhsize + 
    race + 
    weekday + 
    agecat:gender + 
    rolling_pca_3wks + ## mobility data 
    poly(day_std, 3) + ## national polynomial trend 
    (poly(day_std, 3) | state), ## state level pooled polynomial trend 
  family = 'negbinomial',
  inits = 0,
  data = df_formodel_topcode,
  chains = 4, 
  cores = 16,
  iter = 4000)

write_rds(nonhh_model, file = here('data', 'models', 'nonhh_model.rds'))
```

## Additional model specifications 


<!-- Run main model non-hh model with Oxford Stringency Index for paper  -->

<!-- ```{r} -->
<!-- ## mobility + monthly national cubic + state-level hierarchical cubic term  -->
<!-- ## Total execution time: 5346.0 seconds.† -->
<!-- nonhh_model_stringency <- brm( -->
<!--   formula = mvbind(alter_age_0_18, alter_age_18_25, alter_age_25_35, alter_age_35_45, alter_age_45_55, alter_age_55_65, alter_age_65_100) ~  -->
<!--     agecat + -->
<!--     gender + -->
<!--     hhsize +  -->
<!--     race +  -->
<!--     weekday +  -->
<!--     agecat:gender +  -->
<!--     stringency_index + -->
<!--     rolling_pca_3wks + -->
<!--     poly(day_std, 3) + ## national polynomial trend  -->
<!--     (poly(day_std, 3) | state), ## state level pooled polynomial trend  -->
<!--   family = 'negbinomial', -->
<!--   inits = 0, -->
<!--   data = df_formodel_topcode, -->
<!--   chains = 4, -->
<!--   cores = 16, -->
<!--   iter = 4000) -->

<!-- write_rds(nonhh_model_stringency, file = here('data', 'models', 'nonhh_model_stringency.rds')) -->
<!-- ``` -->

<!-- Run main model non-hh model with Oxford Stringency Index (no gov response data) -->

<!-- ```{r} -->
<!-- ## mobility + monthly national cubic + state-level hierarchical cubic term  -->
<!-- ## Total execution time: 5346.0 seconds.† -->
<!-- nonhh_model_stringency_no_mobility <- brm( -->
<!--   formula = mvbind(alter_age_0_18, alter_age_18_25, alter_age_25_35, alter_age_35_45, alter_age_45_55, alter_age_55_65, alter_age_65_100) ~  -->
<!--     agecat + -->
<!--     gender + -->
<!--     hhsize +  -->
<!--     race +  -->
<!--     weekday +  -->
<!--     agecat:gender +  -->
<!--     stringency_index + -->
<!--    #  rolling_pca_3wks + ## government response data  -->
<!--     poly(day_std, 3) + ## national polynomial trend  -->
<!--     (poly(day_std, 3) | state), ## state level pooled polynomial trend  -->
<!--   family = 'negbinomial', -->
<!--   inits = 0, -->
<!--   data = df_formodel_topcode, -->
<!--   chains = 4, -->
<!--   cores = 16, -->
<!--   iter = 4000) -->

<!-- write_rds(nonhh_model_stringency_no_mobility, file = here('data', 'models', 'nonhh_model_stringency_no_mobility.rds')) -->
<!-- ``` -->



<!-- Run main model non-hh without mobility data  -->

<!-- ```{r} -->
<!-- ## mobility + monthly national cubic + state-level hierarchical cubic term  -->
<!-- ## Total execution time: 5346.0 seconds. -->
<!-- nonhh_model_no_mobility <- brm( -->
<!--   formula = mvbind(alter_age_0_18, alter_age_18_25, alter_age_25_35, alter_age_35_45, alter_age_45_55, alter_age_55_65, alter_age_65_100) ~  -->
<!--     agecat + -->
<!--     gender + -->
<!--     hhsize +  -->
<!--     race +  -->
<!--     weekday +  -->
<!--     agecat:gender +  -->
<!--     poly(day_std, 3) + ## national polynomial trend  -->
<!--     (poly(day_std, 3) | state), ## state level pooled polynomial trend  -->
<!--   family = 'negbinomial', -->
<!--   inits = 0, -->
<!--   data = df_formodel_topcode, -->
<!--   chains = 4, -->
<!--   cores = 16, -->
<!--   iter = 4000) -->

<!-- write_rds(nonhh_model_no_mobility, file = here('data', 'models', 'nonhh_model_no_mobility.rds')) -->
<!-- ``` -->
